+++ CHANGELOG.md
@@ -1,14 +1,21 @@
-# [](https://github.com/Evangelos-Komnis-Thesis-ICT-IHU-2025/central-docker-infrastructure/compare/v1.0.0...v) (2026-02-28)
+# [](https://github.com/Evangelos-Komnis-Thesis-ICT-IHU-2025/central-docker-infrastructure/compare/v1.0.0...v) (2026-03-01)
 
 
 ### Bug Fixes
 
+* **ci:** add mock integration mode for ([34bd0d1](https://github.com/Evangelos-Komnis-Thesis-ICT-IHU-2025/central-docker-infrastructure/commit/34bd0d180f89ae4718d923830912cde4f413fa6f))
+* **ci:** allow successful ([73c9a43](https://github.com/Evangelos-Komnis-Thesis-ICT-IHU-2025/central-docker-infrastructure/commit/73c9a43f2c000e3bd466690dd623e8c2430491da))
+* **ci:** checkout dependent repos and ([a5f896a](https://github.com/Evangelos-Komnis-Thesis-ICT-IHU-2025/central-docker-infrastructure/commit/a5f896a3b78f1704506a1bb9de866ce3119222ba))
+* **ci:** harden smoke checks with internal ([e68462f](https://github.com/Evangelos-Komnis-Thesis-ICT-IHU-2025/central-docker-infrastructure/commit/e68462f5bc0ff3f8aa90fae5805203500626afe6))
+* **ci:** make compose integration ([237d29c](https://github.com/Evangelos-Komnis-Thesis-ICT-IHU-2025/central-docker-infrastructure/commit/237d29c501bc141f1f6313b61e227c505f19d7b2))
+* **ci:** skip image builds in mock integration ([f90a617](https://github.com/Evangelos-Komnis-Thesis-ICT-IHU-2025/central-docker-infrastructure/commit/f90a6177f1f74af82728fa2c1679fd7a8dae3314))
 * fix ([a7dc8e6](https://github.com/Evangelos-Komnis-Thesis-ICT-IHU-2025/central-docker-infrastructure/commit/a7dc8e65b5e01bc3c32025a8ad29cb704017c201))
 
 
 ### Features
 
 * **infrastructure:** add Docker Compose ([34b80a9](https://github.com/Evangelos-Komnis-Thesis-ICT-IHU-2025/central-docker-infrastructure/commit/34b80a92b5e8530b6433904ef326d340a7317cbd))
+* **infrastructure:** initialize central ([1b697d8](https://github.com/Evangelos-Komnis-Thesis-ICT-IHU-2025/central-docker-infrastructure/commit/1b697d8e029396e2288af5c8a253bccc49cb31de))